Split workloads.ts mock handlers into focused modules
The monolithic workloads.ts file (2600+ lines) mixed workload, service,
app, pod/envoy/ztunnel, and miscellaneous handlers with all shared types
and factory functions. Split into:

- utils.ts: shared types, interfaces, constants, factory functions,
  definition tables, and namespace getters
- services.ts: service list, detail, metrics, and dashboard handlers
- apps.ts: app list, detail, metrics, and dashboard handlers
- pods.ts: envoy config dump, ztunnel config dump, pod logs/logging,
  custom dashboards, grafana/perses, and stats handlers
- workloads.ts: workload list, detail, metrics, and dashboard handlers
