+++ frontend/src/mocks/handlers.ts
@@ -1,4 +1,5 @@
 // Combine all mock handlers
+import { appHandlers } from './handlers/apps';
 import { authHandlers } from './handlers/auth';
 import { chatbotHandlers } from './handlers/chatbot/chatbot';
 import { configHandlers } from './handlers/config';
@@ -7,12 +8,15 @@ import { istioHandlers } from './handlers/istio';
 import { meshHandlers } from './handlers/mesh';
 import { namespaceHandlers } from './handlers/namespaces';
 import { overviewHandlers } from './handlers/overview';
+import { podHandlers } from './handlers/pods';
+import { serviceHandlers } from './handlers/services';
 import { statusHandlers } from './handlers/status';
 import { tracingHandlers } from './handlers/tracing';
 import { trafficHandlers } from './handlers/traffic';
 import { workloadHandlers } from './handlers/workloads';
 
 export const handlers = [
+  ...appHandlers,
   ...authHandlers,
   ...chatbotHandlers,
   ...configHandlers,
@@ -21,6 +25,8 @@ export const handlers = [
   ...meshHandlers,
   ...namespaceHandlers,
   ...overviewHandlers,
+  ...podHandlers,
+  ...serviceHandlers,
   ...statusHandlers,
   ...tracingHandlers,
   ...trafficHandlers,
